requirements.txt
allure_pytest==2.9.45
allure_python_commons==2.9.45
pytest
requests
